+++ src/hooks/useKeyPress.js
@@ -1,16 +1,29 @@
-import { useEffect } from 'react';
+import { useState, useEffect } from 'react';
 
-export default function useKeyPress(targetKey, callback) {
+export default function useKeyPress(targetKey, callback = () => {}) {
+  const [keyPressed, setKeyPressed] = useState(false);
+
   function downHandler({ key }) {
     if (key === targetKey) {
+      setKeyPressed(true);
       callback();
     }
   }
 
+  const upHandler = ({ key }) => {
+    if (key === targetKey) {
+      setKeyPressed(false);
+    }
+  };
+
   useEffect(() => {
     window.addEventListener('keydown', downHandler);
+    window.addEventListener('keyup', upHandler);
     return () => {
       window.removeEventListener('keydown', downHandler);
+      window.removeEventListener('keyup', upHandler);
     };
   }, []);
+
+  return keyPressed;
 }
